semaphore, tinythread, pthreads SKYEMULATOR IdleLoop

+++ corelib/pthreads/pthread_attr_getstacksize.c
@@ -43,7 +43,7 @@
 #endif
 
 int
-pthread_attr_getstacksize (const pthread_attr_t * attr, size_t * stacksize)
+pthread_attr_getstacksize (const pthread_attr_t * attr, int * stacksize)
      /*
       * ------------------------------------------------------
       * DOCPUBLIC
