+++ corelib/pthreads/pthread_attr_setstacksize.c
@@ -39,7 +39,7 @@
 
 
 int
-pthread_attr_setstacksize (pthread_attr_t * attr, size_t stacksize)
+pthread_attr_setstacksize (pthread_attr_t * attr, int stacksize)
      /*
       * ------------------------------------------------------
       * DOCPUBLIC
